JD Final/requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.31.0
pydantic>=2.0.0
tweepy>=4.12.0
google-cloud-talent>=2.7.0
python-dotenv-vault>=0.5.0
markdown>=3.4.3
google-generativeai>=0.3.0